feat: liste messages bo fusager 587-588
