+++ packages/frontend-bo/src/components/demandes-sejour/Messagerie.vue
@@ -1,0 +1,1 @@
+<template>Messagerie</template>
