+++ packages/frontend-usagers/src/pages/messagerie/index.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="fr-container">
+    <DemandeSejourListe :display="'Messagerie'" />
+  </div>
+</template>
+
+<script setup>
+definePageMeta({
+  middleware: ["is-connected", "check-organisme-is-complet"],
+});
+</script>
