+++ packages/shared/src/utils/index.js
@@ -2,5 +2,6 @@ import createLogger from "./createLogger";
 import hebergement from "./hebergement";
 import regex from "./hebergement";
 import eigUtils from "./hebergement";
+import messageUtils from "./messageUtils";
 
-export { createLogger, hebergement, regex, eigUtils };
+export { createLogger, hebergement, regex, eigUtils, messageUtils };
