+++ packages/shared/src/utils/messageUtils.js
@@ -1,0 +1,9 @@
+const etat = {
+  NON_LU: "NON LU",
+  LU: "LU",
+  REPONDU: "REPONDU",
+};
+
+export default {
+  etat,
+};
